+++ frontend_tests/node_tests/compose_validate.js
@@ -16,10 +16,11 @@ const compose_state = zrequire("compose_state");
 const reminder = mock_esm("../../static/js/reminder");
 const ui_util = mock_esm("../../static/js/ui_util");
 
-const compose = zrequire("compose");
 const compose_pm_pill = zrequire("compose_pm_pill");
+const compose_validate = zrequire("compose_validate");
 const peer_data = zrequire("peer_data");
 const people = zrequire("people");
+const settings_config = zrequire("settings_config");
 const stream_data = zrequire("stream_data");
 
 const me = {
@@ -62,12 +63,12 @@ test_ui("validate_stream_message_address_info", ({mock_template}) => {
         subscribed: true,
     };
     stream_data.add_sub(sub);
-    assert.ok(compose.validate_stream_message_address_info("social"));
+    assert.ok(compose_validate.validate_stream_message_address_info("social"));
 
     sub.subscribed = false;
     stream_data.add_sub(sub);
     mock_template("compose_not_subscribed.hbs", false, () => "compose_not_subscribed_stub");
-    assert.ok(!compose.validate_stream_message_address_info("social"));
+    assert.ok(!compose_validate.validate_stream_message_address_info("social"));
     assert.equal($("#compose-error-msg").html(), "compose_not_subscribed_stub");
 
     page_params.narrow_stream = false;
@@ -76,7 +77,7 @@ test_ui("validate_stream_message_address_info", ({mock_template}) => {
         payload.data.subscribed = true;
         payload.success(payload.data);
     };
-    assert.ok(compose.validate_stream_message_address_info("social"));
+    assert.ok(compose_validate.validate_stream_message_address_info("social"));
 
     sub.name = "Frontend";
     sub.stream_id = 102;
@@ -86,14 +87,14 @@ test_ui("validate_stream_message_address_info", ({mock_template}) => {
         payload.data.subscribed = false;
         payload.success(payload.data);
     };
-    assert.ok(!compose.validate_stream_message_address_info("Frontend"));
+    assert.ok(!compose_validate.validate_stream_message_address_info("Frontend"));
     assert.equal($("#compose-error-msg").html(), "compose_not_subscribed_stub");
 
     channel.post = (payload) => {
         assert.equal(payload.data.stream, "Frontend");
         payload.error({status: 404});
     };
-    assert.ok(!compose.validate_stream_message_address_info("Frontend"));
+    assert.ok(!compose_validate.validate_stream_message_address_info("Frontend"));
     assert.equal(
         $("#compose-error-msg").html(),
         "translated HTML: <p>The stream <b>Frontend</b> does not exist.</p><p>Manage your subscriptions <a href='#streams/all'>on your Streams page</a>.</p>",
@@ -103,7 +104,7 @@ test_ui("validate_stream_message_address_info", ({mock_template}) => {
         assert.equal(payload.data.stream, "social");
         payload.error({status: 500});
     };
-    assert.ok(!compose.validate_stream_message_address_info("social"));
+    assert.ok(!compose_validate.validate_stream_message_address_info("social"));
     assert.equal(
         $("#compose-error-msg").html(),
         $t_html({defaultMessage: "Error checking subscription"}),
@@ -141,7 +142,7 @@ test_ui("validate", ({override, mock_template}) => {
     }
 
     initialize_pm_pill();
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.ok(!$("#sending-indicator").visible());
     assert.ok(!$("#compose-send-button").is_focused());
     assert.equal($("#compose-send-button").prop("disabled"), false);
@@ -151,7 +152,7 @@ test_ui("validate", ({override, mock_template}) => {
     );
 
     reminder.is_deferred_delivery = () => true;
-    compose.validate();
+    compose_validate.validate();
     assert.equal($("#sending-indicator").text(), "translated: Scheduling...");
     reminder.is_deferred_delivery = () => {};
 
@@ -164,7 +165,7 @@ test_ui("validate", ({override, mock_template}) => {
         }
         return false;
     };
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.ok(zephyr_checked);
     assert.equal(
         $("#compose-error-msg").html(),
@@ -180,7 +181,7 @@ test_ui("validate", ({override, mock_template}) => {
     compose_state.set_message_type("private");
 
     compose_state.private_message_recipient("");
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.equal(
         $("#compose-error-msg").html(),
         $t_html({defaultMessage: "Please specify at least one valid recipient"}),
@@ -190,7 +191,7 @@ test_ui("validate", ({override, mock_template}) => {
     add_content_to_compose_box();
     compose_state.private_message_recipient("foo@zulip.com");
 
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
 
     assert.equal(
         $("#compose-error-msg").html(),
@@ -198,7 +199,7 @@ test_ui("validate", ({override, mock_template}) => {
     );
 
     compose_state.private_message_recipient("foo@zulip.com,alice@zulip.com");
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
 
     assert.equal(
         $("#compose-error-msg").html(),
@@ -207,22 +208,22 @@ test_ui("validate", ({override, mock_template}) => {
 
     people.add_active_user(bob);
     compose_state.private_message_recipient("bob@example.com");
-    assert.ok(compose.validate());
+    assert.ok(compose_validate.validate());
 
     people.deactivate(bob);
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.equal(
         $("#compose-error-msg").html(),
         $t_html({defaultMessage: "You cannot send messages to deactivated users."}),
     );
 
     page_params.realm_is_zephyr_mirror_realm = true;
-    assert.ok(compose.validate());
+    assert.ok(compose_validate.validate());
     page_params.realm_is_zephyr_mirror_realm = false;
 
     compose_state.set_message_type("stream");
     compose_state.stream_name("");
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.equal(
         $("#compose-error-msg").html(),
         $t_html({defaultMessage: "Please specify a stream"}),
@@ -231,7 +232,7 @@ test_ui("validate", ({override, mock_template}) => {
     compose_state.stream_name("Denmark");
     page_params.realm_mandatory_topics = true;
     compose_state.topic("");
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.equal(
         $("#compose-error-msg").html(),
         $t_html({defaultMessage: "Please specify a topic"}),
@@ -255,7 +256,59 @@ test_ui("get_invalid_recipient_emails", ({override}) => {
     people.initialize(page_params.user_id, params);
 
     override(compose_state, "private_message_recipient", () => "welcome-bot@example.com");
-    assert.deepEqual(compose.get_invalid_recipient_emails(), []);
+    assert.deepEqual(compose_validate.get_invalid_recipient_emails(), []);
+});
+
+test_ui("test_wildcard_mention_allowed", () => {
+    page_params.user_id = me.user_id;
+
+    page_params.realm_wildcard_mention_policy =
+        settings_config.wildcard_mention_policy_values.by_everyone.code;
+    page_params.is_guest = true;
+    page_params.is_admin = false;
+    assert.ok(compose_validate.wildcard_mention_allowed());
+
+    page_params.realm_wildcard_mention_policy =
+        settings_config.wildcard_mention_policy_values.nobody.code;
+    page_params.is_admin = true;
+    assert.ok(!compose_validate.wildcard_mention_allowed());
+
+    page_params.realm_wildcard_mention_policy =
+        settings_config.wildcard_mention_policy_values.by_members.code;
+    page_params.is_guest = true;
+    page_params.is_admin = false;
+    assert.ok(!compose_validate.wildcard_mention_allowed());
+
+    page_params.is_guest = false;
+    assert.ok(compose_validate.wildcard_mention_allowed());
+
+    page_params.realm_wildcard_mention_policy =
+        settings_config.wildcard_mention_policy_values.by_moderators_only.code;
+    page_params.is_moderator = false;
+    assert.ok(!compose_validate.wildcard_mention_allowed());
+
+    page_params.is_moderator = true;
+    assert.ok(compose_validate.wildcard_mention_allowed());
+
+    page_params.realm_wildcard_mention_policy =
+        settings_config.wildcard_mention_policy_values.by_stream_admins_only.code;
+    page_params.is_admin = false;
+    assert.ok(!compose_validate.wildcard_mention_allowed());
+
+    // TODO: Add a by_admins_only case when we implement stream-level administrators.
+
+    page_params.is_admin = true;
+    assert.ok(compose_validate.wildcard_mention_allowed());
+
+    page_params.realm_wildcard_mention_policy =
+        settings_config.wildcard_mention_policy_values.by_full_members.code;
+    const person = people.get_by_user_id(page_params.user_id);
+    person.date_joined = new Date(Date.now());
+    page_params.realm_waiting_period_threshold = 10;
+
+    assert.ok(compose_validate.wildcard_mention_allowed());
+    page_params.is_admin = false;
+    assert.ok(!compose_validate.wildcard_mention_allowed());
 });
 
 test_ui("validate_stream_message", ({override, mock_template}) => {
@@ -274,7 +327,7 @@ test_ui("validate_stream_message", ({override, mock_template}) => {
     };
     stream_data.add_sub(sub);
     compose_state.stream_name("social");
-    assert.ok(compose.validate());
+    assert.ok(compose_validate.validate());
     assert.ok(!$("#compose-all-everyone").visible());
     assert.ok(!$("#compose-send-status").visible());
 
@@ -291,16 +344,16 @@ test_ui("validate_stream_message", ({override, mock_template}) => {
         compose_content = data;
     };
 
-    override(compose, "wildcard_mention_allowed", () => true);
+    override(compose_validate, "wildcard_mention_allowed", () => true);
     compose_state.message_content("Hey @**all**");
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.equal($("#compose-send-button").prop("disabled"), false);
     assert.ok(!$("#compose-send-status").visible());
     assert.equal(compose_content, "compose_all_everyone_stub");
     assert.ok($("#compose-all-everyone").visible());
 
-    override(compose, "wildcard_mention_allowed", () => false);
-    assert.ok(!compose.validate());
+    override(compose_validate, "wildcard_mention_allowed", () => false);
+    assert.ok(!compose_validate.validate());
     assert.equal(
         $("#compose-error-msg").html(),
         $t_html({
@@ -327,7 +380,7 @@ test_ui("test_validate_stream_message_post_policy_admin_only", ({override}) => {
     compose_state.topic("subject102");
     compose_state.stream_name("stream102");
     stream_data.add_sub(sub);
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.equal(
         $("#compose-error-msg").html(),
         $t_html({defaultMessage: "Only organization admins are allowed to post to this stream."}),
@@ -341,7 +394,7 @@ test_ui("test_validate_stream_message_post_policy_admin_only", ({override}) => {
 
     compose_state.topic("subject102");
     compose_state.stream_name("stream102");
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.equal(
         $("#compose-error-msg").html(),
         $t_html({defaultMessage: "Only organization admins are allowed to post to this stream."}),
@@ -365,7 +418,7 @@ test_ui("test_validate_stream_message_post_policy_moderators_only", ({override})
     compose_state.topic("subject104");
     compose_state.stream_name("stream104");
     stream_data.add_sub(sub);
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.equal(
         $("#compose-error-msg").html(),
         $t_html({
@@ -402,7 +455,7 @@ test_ui("test_validate_stream_message_post_policy_full_members_only", ({override
     compose_state.topic("subject103");
     compose_state.stream_name("stream103");
     stream_data.add_sub(sub);
-    assert.ok(!compose.validate());
+    assert.ok(!compose_validate.validate());
     assert.equal(
         $("#compose-error-msg").html(),
         $t_html({defaultMessage: "Guests are not allowed to post to this stream."}),
